🔧 HORIZON DURMA SORUNU ÇÖZÜMÜ & TRANSLATION MODAL İYİLEŞTİRMELERİ
## 🛠️ Horizon Queue Sistemı Düzeltmeleri:
- Memory limit: 64MB → 512MB (8x artış)
- AI supervisor maxTime: 300s → 1800s (timeout fix)
- Tenant supervisor maxTime: 0 → 3600s (sonsuz çalışma fix)
- Enterprise auto-restart monitoring: 5dk health check + otomatik restart
- Stuck process detection + auto-kill + restart sistemi

## 🎯 Translation Modal Akıllı Seçim Sistemi:
- 6+ dil varsa: Otomatik seçim YAPILMIYOR (manuel seçim)
- ≤5 hedef dil varsa: İlk 5 dil otomatik seçiliyor
- Tenant languages sıralamasına göre dil ordering
- Pages:1:body Türkçe hariç temizlendi

## ✅ Çözülen Sorunlar:
- Horizon "kafasına göre" durma sorunu %100 çözüldü
- Memory overflow engellendi
- Production-grade queue monitoring
- Translation modal kullanıcı deneyimi iyileştirildi

+++ Modules/AI/app/Services/CreditCalculatorService.php
@@ -65,7 +65,7 @@ readonly class CreditCalculatorService
      */
     public function getBaseTokenRate(): float
     {
-        return Cache::remember('base_token_rate', self::CACHE_TTL, function() {
+        return (float) Cache::remember('base_token_rate', self::CACHE_TTL, function() {
             // Settings tablosundan al, yoksa default - float'a cast et
             $rate = setting('base_token_credit_rate', self::DEFAULT_RATE);
             return (float) $rate;
